meh ---- Start refactoring factory code out into submodules. Let's see if we can separate class-based plugin factories and configurable plugin factories
