+++ betty/datas/plugin/definition/event_type.py
@@ -22,7 +22,7 @@ from betty.sample import Sample
     EventTypeDefinition,
     samples=[
         lambda: Sample(
-            EventTypeDefinitionData(
+            EventTypeDefinitionConfig(
                 id="moon-landing",
                 label="Moon landing",
                 label_plural="Moon landings",
@@ -37,7 +37,7 @@ from betty.sample import Sample
         )
     ],
 )
-class EventTypeDefinitionData(
+class EventTypeDefinitionConfig(
     CountableHumanFacingPluginDefinitionData[EventTypeDefinition],
     OrderedPluginDefinitionData[EventTypeDefinition],
 ):
