+++ betty/tests/datas/plugin/definition/test_role.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from betty.datas.plugin.definition.role import RoleDefinitionData
+from betty.datas.plugin.definition.role import RoleDefinitionConfig
 from betty.localizables.plain import Plain
 from betty.test_utils.locale.localizable import (
     DUMMY_COUNTABLE_LOCALIZABLE,
@@ -12,7 +12,7 @@ class TestRoleDefinitionData:
         plugin_id = "my-first-role"
         label = Plain("-")
         label_plural = Plain("-")
-        sut = RoleDefinitionData(
+        sut = RoleDefinitionConfig(
             id=plugin_id,
             label=label,
             label_plural=label_plural,
@@ -26,7 +26,7 @@ class TestRoleDefinitionData:
 
     def test_new_plugin__full(self) -> None:
         description = Plain("-")
-        sut = RoleDefinitionData(
+        sut = RoleDefinitionConfig(
             id="my-first-role",
             label="-",
             label_plural="-",
